Update detection-board-marbles.js
 color thresholds for black and white balls are defined and used to classify the circles based on their color.

The color of the pixel at the center of each circle is extracted using the img.ucharPtr() method. If the pixel value is below the threshold for black balls, the circle is considered a black ball. If the pixel value is above the threshold for white balls, the circle is considered a white ball. Otherwise, the circle is considered an empty space.

Note that the threshold values for black and white balls may vary depending on the lighting and image quality. You may need to adjust these values to get the best results.