Create service-worker.js
We used a constant to store the cache name and added a version number to ensure it is unique. We also added CSS, JS, image, and font files to the list of URLs to be cached. Finally, we used a listener to intercept network requests and return the cached response if it exists, otherwise, perform the request normally.